The origin of this commit was attempting to write Filters that require access to the database to perform their work. I wanted to use DI as I had throughout the rest of the project. The DI version required most services including EfGraphQLService to be scoped. The default configuration recommendation using EfGraphQLConventions registers these services as Singletons.
I modified EfGraphQLService to contain a constructor that accepts Filters directly. The same is true for the TDbContext. This change permits a DI configuration that looks like this:

```bash
            services.AddScoped<ZigletDbContext>();
            services.AddScoped<Mutation.Mutation>();
            services.AddScoped<Query.Query>();
            services.AddScoped<ISchema, Schema.Schema>();
            services.AddHttpContextAccessor();
            services.AddScoped<Filters>();
            services.AddScoped<UserFilter>();
            services.AddScoped<HttpContextAccessor>();
            EfGraphQLConventions.RegisterScalarsAndArgs(services);
            services.AddScoped<IEfGraphQLService<ZigletDbContext>,EfGraphQLService<ZigletDbContext>>();
```

This implementation alleviates confusion where some parts of the application expect to receive injected services where other parts of the application do their own service resolution like :

```bash
  using(var scope = _provider.CreateScope()){
    var service = scope.GetService<TService>();
  }
```

I recommend that in general the system be simplified so that all classes are unaware of the DI container and that none take parameters of types [`IServiceProvider`,`IServiceCollection`,`IServiceScopeFactory`].